// resources/js/modules/manufacturer/redux/manufacturer.types.ts
enum ManufacturerTypes {
    MANUFACTURER_ACTION_START = "MANUFACTURER_ACTION_START",
    MANUFACTURER_GET_SUCCESS = "MANUFACTURER_GET_SUCCESS",
    MANUFACTURER_GET_FAILED = "MANUFACTURER_GET_FAILED",
    MANUFACTURER_CREATE_SUCCESS = "MANUFACTURER_CREATE_SUCCESS",
    MANUFACTURER_CREATE_FAILED = "MANUFACTURER_CREATE_FAILED",
    MANUFACTURER_GET_BY_ID_SUCCESS = "MANUFACTURER_GET_BY_ID_SUCCESS",
    MANUFACTURER_GET_BY_ID_FAILED = "MANUFACTURER_GET_BY_ID_FAILED",
    MANUFACTURER_UPDATE_SUCCESS = "MANUFACTURER_UPDATE_SUCCESS",
    MANUFACTURER_UPDATE_FAILED = "MANUFACTURER_UPDATE_FAILED",
    MANUFACTURER_DELETE_SUCCESS = "MANUFACTURER_DELETE_SUCCESS",
    MANUFACTURER_DELETE_FAILED = "MANUFACTURER_DELETE_FAILED",
    MANUFACTURER_ACTION_END = "MANUFACTURER_ACTION_END",
}

export default ManufacturerTypes;